// src/api/__generated__/story/model/dramaListItemResponse.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Mini Drama Service API
 * Mini Drama Platform
 * OpenAPI spec version: v1.0
 */
import type { ActorCollectionInfoResponse } from './actorCollectionInfoResponse';
import type { DramaListItemResponseBadge } from './dramaListItemResponseBadge';

/**
 * 数据列表
 */
export interface DramaListItemResponse {
  dramaId?: number;
  dramaTitle?: string;
  dramaDescription?: string;
  dramaCoverUrl?: string;
  tags?: string[];
  creatorName?: string;
  /** Drama creator avatar URL */
  creatorAvatarUrl?: string;
  creatorId?: number;
  badge?: DramaListItemResponseBadge;
  avgRating?: number;
  totalEpisodes?: number;
  totalPlayCount?: number;
  totalCompletedViewCount?: number;
  totalHeatValue?: number;
  actorCollections?: ActorCollectionInfoResponse[];
}